requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bibtexparser==1.4.1
blinker==1.8.1
browser-cookie3==0.19.1
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
configparser==6.0.1
contourpy==1.2.1
coverage==7.5.3
cryptography==42.0.5
cycler==0.12.1
DateTime==5.5
debugpy==1.8.1
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
duckdb==0.10.2
email_validator==2.1.2
et-xmlfile==1.1.0
etelemetry==0.3.1
executing==2.0.1
extra-streamlit-components==0.1.71
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.20.0
feedparser==6.0.11
filelock==3.13.3
fitz==0.0.1.dev2
Flask==2.1.3
Flask-Login==0.4.1
fonttools==4.53.0
frontend==0.0.3
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.1
googletrans==3.0.0
gotrue==2.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
html2text==2024.2.26
html5lib==1.1
httpcore==0.9.1
httplib2==0.22.0
httptools==0.6.1
httpx==0.13.3
huggingface-hub==0.22.2
hyperframe==5.2.0
icecream==2.1.3
idna==2.10
iniconfig==2.0.0
install==1.3.5
ipykernel==6.29.4
ipython==8.23.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpickle==3.0.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
loguru==0.7.2
looseversion==1.3.0
lxml==5.2.1
lz4==4.3.3
macos-notifications==0.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nbclient==0.10.0
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.0.1
openai==1.33.0
openpyxl==3.1.2
orjson==3.10.5
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
paper-qa==4.9.0
parso==0.8.3
pathlib==1.0.1
pendulum==3.0.0
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.5.0
postgrest==0.16.8
progressbar==2.5
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.25.3
prov==2.0.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pushover==0.5
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
pydot==2.0.0
PyDrive==1.3.1
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyobjc-core==10.2
pyobjc-framework-Cocoa==10.2
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
pyperclip==1.8.2
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytils==0.4.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.2
pyzotero==1.5.19
rauth==0.7.3
rdflib==6.3.2
realtime==1.0.6
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.13.0
selenium==4.9.0
sentence-transformers==2.6.1
setuptools==69.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
st-login-form==0.2.3
stack-data==0.6.3
starlette==0.37.2
storage3==0.7.6
streamlit==1.35.0
streamlit-authenticator==0.3.2
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
sympy==1.12
tenacity==8.4.1
threadpoolctl==3.4.0
tiktoken==0.7.0
time-machine==2.14.1
tokenizers==0.15.2
toml==0.10.2
tools==0.1.9
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
traits==6.3.2
transformers==4.39.3
treelib==1.7.0
trio==0.25.0
trio-websocket==0.11.1
typer==0.12.3
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==2.0.3
wsproto==1.2.0
zope.interface==6.2
